+++ CHANGELOG.md
@@ -66,3 +66,4 @@
 
 -   CV acquisition at _1/16x_ sample rate
 -   support for dark mode (the NES display no longer dims)
+-   fix logic for changing the sample rate